requirements.txt
asgiref==3.4.1
astroid==2.7.3
bandit==1.7.0
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
gitdb==4.0.7
GitPython==3.1.18
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
pbr==5.6.0
platformdirs==2.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pylint==2.10.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
stevedore==3.4.0
toml==0.10.2
wrapt==1.12.1
